Fix rspec issues
