--- app/graphql/models/photo.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  class Photo < Types::BaseObject
-    field :id, GraphQL::Types::ID, null: false
-    field :uri, Types::Url, null: false
-  end
-end
