+++ app/graphql/types/models/project.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Types
+  module Models
+    class Project < Types::BaseObject
+      field :id, GraphQL::Types::ID, null: false
+      field :name, String, null: false
+      field :description, String, null: false
+      field :photo, Models::Photo, null: false
+      field :technologies, [Models::Technology], null: false
+    end
+  end
+end
